Implement 'info' and 'delete' commands

+++ script/repotoollib/gitlab.py
@@ -10,11 +10,14 @@ from repotoollib.util import make_url
 
 _GITLAB_API_URL = "https://gitlab.com/api/v4"
 
-def _encode_project_name_or_id(user, name_or_id):
-    if isinstance(name_or_id, int):
-        return str(name_or_id)
+def _encode_project_name(user, project_name):
+    return urllib.quote_plus("{}/{}".format(user, project_name))
+
+def _encode_project_name_or_id(user, project_name_or_id):
+    if isinstance(project_name_or_id, int):
+        return str(project_name_or_id)
     else:
-        return urllib.quote_plus("{}/{}".format(user, name_or_id))
+        return _encode_project_name(user, project_name_or_id)
 
 def _make_project(provider, project_obj):
     clone_links = {
@@ -48,44 +51,63 @@ class GitLab(object):
 
         projects = []
         while True:
-            url = make_url(
+            r = self._do_request(
+                "get",
                 _GITLAB_API_URL,
                 "users",
                 self._user,
                 "projects",
                 query)
-            r = requests.get(url)
-            r.raise_for_status()
-
             projects.extend(map(lambda o: _make_project(self, o), r.json()))
-
             page_id = r.headers.get("X-Next-Page")
             if page_id is None or len(page_id) == 0: break
             query["page"] = page_id
 
         return projects
 
-    def create_project(self, name, visibility="private"):
+    def project(self, project_name):
+        r = self._do_request(
+            "get",
+            _GITLAB_API_URL,
+            "projects",
+            self._encode_project_name(project_name),
+            private_token=self._api_token)
+        return _make_project(self, r.json())
+
+    def create_project(self, project_name, visibility="private"):
         url = make_url(
             _GITLAB_API_URL,
             "projects",
             private_token=self._api_token)
-        r = requests.post(url, data={ "name": name, "visibility": visibility })
+        r = requests.post(url, data={ "name": project_name, "visibility": visibility })
         r.raise_for_status()
 
-    def delete_project(self, name_or_id):
+    def delete_project(self, project_name, confirmation_token=False):
+        if not confirmation_token:
+            raise RuntimeError("Dangerous operation disallowed")
+
         url = make_url(
             _GITLAB_API_URL,
             "projects",
-            _encode_project_name_or_id(self._user, name_or_id))
+            _encode_project_name_or_id(self._user, project_name))
         r = requests.delete(url, data={ "private_token": self._api_token })
         r.raise_for_status()
 
-    def archive_project(self, name_or_id):
+    def archive_project(self, project_name):
+        raise RuntimeError("Not implemented")
         url = make_url(
             _GITLAB_API_URL,
             "projects",
-            _encode_project_name_or_id(self._user, name_or_id),
+            _encode_project_name_or_id(self._user, project_name),
             "archive")
         r = requests.post(url, data={ "private_token": self._api_token })
         r.raise_for_status()
+
+    def _encode_project_name(self, project_name):
+        return urllib.quote_plus("{}/{}".format(self._user, project_name))
+
+    def _do_request(self, method, *args, **kwargs):
+        url = make_url(*args, **kwargs)
+        r = requests.request(method, url)
+        r.raise_for_status()
+        return r
